+++ src/lib/main/fetchDiaries.ts
@@ -1,9 +1,8 @@
 "use client";
 import { SortedDiaries } from "@/types/main/Calendar";
 import { useInfiniteQuery, useQuery } from "@tanstack/react-query";
-// import { useInfiniteQuery} from "@tanstack/react-query";
 import browserClient from "@/utils/supabase/client";
-
+//=====================================================
 //NOTE - 처음 일기 데이터 받아오기
 export const getInitialDiaries = async () => {
   const { data: initialDiaries } = await browserClient.from("diary").select().order("date", { ascending: false });
@@ -20,7 +19,7 @@ export const useFetchDiaries = () => {
   });
   return { data, error, isLoading };
 };
-
+//=====================================================
 //NOTE - 조회한 날짜 일기 데이터 가져오기
 export const getSelectedDiaries = async (startDate: string, endDate: string) => {
   const { data: selectedDiaries } = await browserClient
@@ -32,19 +31,52 @@ export const getSelectedDiaries = async (startDate: string, endDate: string) => 
 
   return selectedDiaries as SortedDiaries[];
 };
-
-//NOTE -
+//=====================================================
+//NOTE - 검색한 단어가 들어있는 데이터 가져오기
 export const getSearchDiaries = async (value: string) => {
   if (value) {
     const { data: searchDiaries } = await browserClient.from("diary").select().like("contents", `%${value}%`);
     return searchDiaries as SortedDiaries[];
   }
 };
+//=====================================================
+//NOTE - 검색한 단어가 들어있는 데이터 페이지 단위로 가져오기
+// export const getSearchPaginatedDiaries = async (pageParam: number, limit: number, value: string) => {
+//   const from = (pageParam - 1) * limit;
+//   const to = pageParam * limit - 1;
+//   const {
+//     data: searchPaginatedDiaries,
+//     error,
+//     count
+//   } = await browserClient
+//     .from("diary")
+//     .select("*", { count: "exact" })
+//     .like("contents", `%${value}%`)
+//     .order("date", { ascending: false })
+//     .range(from, to);
+//   if (error) {
+//     throw new Error(error.message);
+//   }
+//   const hasNext = to + 1 < (count || 0);
+//   return { searchPaginatedDiaries, hasNext, nextPage: pageParam + 1, count };
+// };
+// //NOTE - useInfiniteQuery 사용하기
+// export const useInfiniteQuerySearchDiaries = (query) => {
+//   const { data, isError, isLoading, fetchNextPage, hasNextPage, isFetchingNextPage } = useInfiniteQuery({
+//     queryKey: ["searchDiaries", query],
+//     initialPageParam: 1,
+//     queryFn: ({ pageParam, query }) => getSearchPaginatedDiaries(pageParam, 5, query),
+//     getNextPageParam: (lastPage) => {
+//       return lastPage?.hasNext ? lastPage.nextPage : undefined;
+//     }
+//   });
 
+//   return { data, isError, isLoading, fetchNextPage, hasNextPage, isFetchingNextPage };
+// };
+
+//=====================================================
 //NOTE - 페이지 단위로 데이터 가져오기
 export const getPaginatedDiaries = async (pageParam: number, limit: number) => {
-  console.log("pageParam", pageParam);
-
   const from = (pageParam - 1) * limit; //10
   const to = pageParam * limit - 1; //19
